requirements.txt
# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Processing
pyyaml>=6.0
pydantic>=2.0.0

# TACO Dataset Tools
pycocotools>=2.0.7
requests>=2.31.0
tqdm>=4.66.0
